Add files via upload
In this game we play by pressing the color button and while playing always remember its pattern like press the color which is blinking and then if any other button is blinking then follow the pattern from the beginning before pressing it. and always follow the pattern.so that you can move towards higher scores.